[FreeBSD] move FreeBSDGNUReplacements to a freebsd folder
