+++ xbmc/addons/Addon.cpp
@@ -30,7 +30,7 @@
 #include "../osx/OSXGNUReplacements.h"
 #endif
 #ifdef __FreeBSD__
-#include "FreeBSDGNUReplacements.h"
+#include "freebsd/FreeBSDGNUReplacements.h"
 #endif
 #include "utils/log.h"
 #include "utils/URIUtils.h"
