+++ xbmc/filesystem/SAPDirectory.cpp
@@ -32,7 +32,7 @@
 #include "OSXGNUReplacements.h" // strnlen
 #endif
 #ifdef __FreeBSD__
-#include "FreeBSDGNUReplacements.h"
+#include "freebsd/FreeBSDGNUReplacements.h"
 #endif
 
 #include <sys/socket.h>
