+++ xbmc/freebsd/Makefile.in
@@ -1,0 +1,11 @@
+ARCH=@ARCH@
+
+CXXFLAGS+=-fPIC
+
+SRCS=FreeBSDGNUReplacements.c
+
+LIB=freebsd.a
+
+include @abs_top_srcdir@/Makefile.include
+
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
